implemented matching RecordNode function

+++ src/common/Item.java
@@ -5,34 +5,34 @@ package common;
  * @since 14.04.16.
  */
 public class Item implements Comparable<Item> {
-    public float[] values;
+    public double[] values;
     public String name;
 
     public Item() {
     }
 
-    public Item(float[] values, String name) {
+    public Item(double[] values, String name) {
         this.values = values;
         this.name = name;
     }
 
-    public Item(float[] values) {
+    public Item(double[] values) {
         this(values, null);
     }
 
     @Override
     public int compareTo(Item o) {
-        return Float.compare(getMetric(this), getMetric(o));
+        return Double.compare(getMetric(this), getMetric(o));
     }
 
 
-    public static float getMetric(Item item) {
-        float sum = 0f;
-        float[] values = item.values;
-        for (float value : values) {
+    public static double getMetric(Item item) {
+        double sum = 0f;
+        double[] values = item.values;
+        for (double value : values) {
             sum += value * value;
         }
-        return (float) Math.sqrt(sum);
+        return   Math.sqrt(sum);
     }
 
     /**
@@ -42,7 +42,7 @@ public class Item implements Comparable<Item> {
      * @return Item considered as difference between two Items
      */
     public static Item diff(Item candidate, Item object) {
-        float[] values = new float[candidate.values.length];
+        double[] values = new double[candidate.values.length];
         for (int j = 0; j < candidate.values.length; j++)
             values[j] = candidate.values[j] - object.values[j];
         return new Item(values);
