+++ src/topics/nearest_k_neighbours/NearestKNeighbours.java
@@ -26,7 +26,7 @@ public class NearestKNeighbours {
         //instantiate generic dataReader with given readStrategy
         DataReader<Item> dataReader = new DataReader<>(readStrategy);
         //create candidate
-        Item candidate = new Item(new float[]{1f, 6f, 3f, 4f});
+        Item candidate = new Item(new double[]{1f, 6f, 3f, 4f});
         //get all
         List<Item> list = dataReader.read("randomizer.txt");
         if (list.size() < NEAREST_CANDIDATES || !DataReader.dataSetOk(list))
